+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## [Eyes.Sdk.Core vNext]
+### Added
+- `UseCookies` configuration option [Trello 2460] (https://trello.com/c/pR7tpNWu)
 ### Fixed
 - Fixed `Eyes.Open` on .NET Framework versions <= 4.8 [Trello 2508] (https://trello.com/c/VYEdbQkM)
+
+## [Eyes.Selenium vNext]
+### Added
+- Ultrafast Grid: Cookies support when downloading resources from another domain [Trello 2460] (https://trello.com/c/pR7tpNWu)
 
 ## [Eyes.Appium 4.20] - 2021-03-18
 ### Updated